% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truth.R
\name{expand_truth_tbl}
\alias{expand_truth_tbl}
\title{Generate a truth table}
\usage{
expand_truth_tbl(fact_names)
}
\arguments{
\item{fact_names}{A character vector of facts.}
}
\value{
A data.frame.
}
\description{
Generate a truth table given a vector of facts.
}
\examples{
expand_truth_tbl(c("is_neutered", "is_inactive"))
}
